// generated-source/models/club-follower.entity-model.ts
import { IUser } from './user.entity-model';
import { IClub } from './club.entity-model';
export interface IClubFollower {
  id: number;
  uuid: string;
  createdAt: Date;
  updatedAt: Date;
  deletedAt: Date;
  createdBy?: string;
  updatedBy?: string;
  deletedBy?: string;
  user?: IUser;
  club?: IClub;
}